Zalando, set shipping done

+++ JSON/configJson.cs
@@ -4,7 +4,7 @@
     {
         public string license_key { get; set; }
         public string captchaKey { get; set; }
-        public int delayBetweenTasks { get; set; }
+        public float delayBetweenTasks { get; set; }
         public string success_webhook { get; set; }
         public string fail_webhook { get; set; }
     }
